+++ ergo/ChangeLog.md
@@ -23,6 +23,7 @@
 * Improve runtime task prioritization.
   * Previously priorities were only used when a task began, but now they are
     used at any await points in tasks.
+* Format markdown output in the terminal when `--doc` is used.
 
 ## 1.0.0-rc.0  -- 2021-08-01
 Most of the changes listed below are breaking changes, so there is no separate